yt-mcp/pyproject.toml
[project]
name = "yt-transcribe"
version = "0.1.0"
description = "MCP server for downloading and transcribing videos using yt-dlp and AssemblyAI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0.0",
    "assemblyai>=0.30.0",
    "python-dotenv>=1.0.0",
]
# Note: yt-dlp must be installed system-wide (e.g., `brew install yt-dlp`)

[project.scripts]
yt-transcribe = "yt_transcribe.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/yt_transcribe"]
